refactor(ui): extract UserMessageDisplay to dedicated module
Slice 11b.1 step 7/14 of the message-part.tsx split. Moves the user-
side legacy rendering surface — the pre-rewrite bubble layout used by
the existing SessionTurn shell — into `user-message-display.tsx`. The
co-located `HighlightedText` helper (the only consumer of file-
reference highlighting in the user body) moves with it.

This is the second of the two named extractions in design doc §1
("extracted from message-part.tsx L1076"). With both `User-` and
`AssistantMessageDisplay` now in their own modules, the path the
W1 sibling slice will take to remove the legacy SessionTurn is one
deletion-only commit per surface.

`message-part.tsx` still re-exports `UserMessageDisplay` so existing
consumers (session-turn.tsx, stories, tests) keep working. The
sibling slice migrates them to import directly from
`./user-message-display`.

message-part.tsx 1607 → 1402 lines (-205). Typecheck GREEN, 478 pass /
3 baseline fail unchanged.
